add spotless and finalize the bean
