--- apps/openchallenges/challenge-service/src/main/java/org/sagebionetworks/openchallenges/challenge/service/configuration/EurekaInstanceConfiguration.java
@@ -1,17 +1,0 @@
-package org.sagebionetworks.openchallenges.challenge.service.configuration;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class EurekaInstanceConfiguration {
-
-  @Bean
-  @Profile("!default")
-  public EurekaInstanceConfigBean eurekaInstanceConfig(InetUtils inetUtils) {
-    EurekaInstanceConfigBean bean = new EurekaInstanceConfigBean(inetUtils);
-    AmazonInfo info = AmazonInfo.Builder.newBuilder().autoBuild("eureka");
-    bean.setDataCenterInfo(info);
-    return bean;
-  }
-}
