+++ apps/openchallenges/service-registry/build.gradle
@@ -5,6 +5,7 @@ buildscript {
 }
 
 plugins {
+  id 'com.diffplug.spotless' version "${spotlessVersion}"
   id 'java'
   id 'org.springframework.boot' version "${springBootVersion}"
   id "io.spring.dependency-management" version "${springDependencyManagementVersion}"
@@ -40,4 +41,14 @@ tasks.withType(JavaCompile) {
 
 bootBuildImage {
   imageName = "ghcr.io/sage-bionetworks/${project.name}-base:local"
+}
+
+spotless {
+  java {
+    target 'src/*/java/**/*.java'
+
+    importOrder()
+    removeUnusedImports()
+    googleJavaFormat()
+  }
 }
