+++ apps/openchallenges/service-registry/gradle.properties
@@ -1,4 +1,5 @@
 openchallengesVersion=0.0.1-SNAPSHOT
+spotlessVersion=6.11.0
 springBootVersion=2.7.8
 springCloudVersion=3.1.4
 springDependencyManagementVersion=1.0.14.RELEASE
