+++ apps/openchallenges/service-registry/src/main/java/org/sagebionetworks/openchallenges/service/registry/ServiceRegistryApplication.java
@@ -30,5 +30,4 @@ public class ServiceRegistryApplication implements CommandLineRunner {
   public void run(String... args) throws Exception {
     LOG.info(serviceRegistryConfigData.getWelcomeMessage());
   }
-
 }
