+++ apps/openchallenges/service-registry/src/test/java/org/sagebionetworks/openchallenges/service/registry/ServiceRegistryApplicationTests.java
@@ -6,8 +6,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class ServiceRegistryApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
-
+  @Test
+  void contextLoads() {}
 }
