+++ libs/openchallenges/app-config-data/src/main/java/org/sagebionetworks/openchallenges/app/config/data/ServiceRegistryConfigData.java
@@ -10,4 +10,5 @@ import org.springframework.context.annotation.Configuration;
 public class ServiceRegistryConfigData {
 
   private String welcomeMessage;
+  private boolean isDeployedOnAws;
 }
